+++ miniconf/tests/arrays.rs
@@ -95,9 +95,9 @@ fn not_found() {
 
 #[test]
 fn metadata() {
-    let metadata = Settings::metadata().separator("/");
+    let metadata = Settings::metadata();
     assert_eq!(metadata.max_depth, 4);
-    assert_eq!(metadata.max_length, "/aam/0/0/c".len());
+    assert_eq!(metadata.max_length("/"), "/aam/0/0/c".len());
     assert_eq!(metadata.count, 11);
 }